\name{EMundirK1-package}
\alias{EMundirK1-package}
\alias{EMundirK1}
\docType{package}
\title{
\packageTitle{EMundirK1}
}
\description{
\packageDescription{EMundirK1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMundirK1}
\packageIndices{EMundirK1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMundirK1}

Maintainer: \packageMaintainer{EMundirK1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
